packages/server/src/auth/guard/jwt.guard.ts
import { Observable } from 'rxjs';

import {
  ExecutionContext,
  Injectable,
  UnauthorizedException
} from '@nestjs/common';
import { AuthGuard } from '@nestjs/passport';

@Injectable()
export class JwtGuard extends AuthGuard('jwt') {
  canActivate(
    context: ExecutionContext
  ): boolean | Promise<boolean> | Observable<boolean> {
    return super.canActivate(context);
  }

  // eslint-disable-next-line @typescript-eslint/no-unused-vars
  handleRequest(error: any, user: any, info: any) {
    if (error || !user) {
      throw error || new UnauthorizedException();
    }

    return user;
  }
}
